Got really annoyed that even with the ridge markers hidden by the store, the Z key was still active, and the markers would fold and dig a trench invisible
Had to make another custom specialization to fix it

I also decided to make the ridge markers fold up whenever it made sense automagically.  Some might not like that.

Using the nifty initSpecialization() function in herb

Added a markerVisibilityNode to the xml to help with detecting if the store has hidden the markers